+++ src/core/Engine.cpp
@@ -6,6 +6,9 @@
 #include "input/InputConfig.h"
 #include "graphics/FontManager.h"
 #include "graphics/Font5x7.h"
+#include "graphics/Color.h"
+#include <cstdio>
+#include <cstring>
 
 namespace pixelroot32::core {
 
@@ -17,18 +20,33 @@ namespace pixelroot32::core {
         : renderer(displayConfig), inputManager(inputConfig), audioEngine(audioConfig), musicPlayer(audioEngine) {
         previousMillis = 0;
         deltaTime = 0;
+#ifdef PIXELROOT32_ENABLE_FPS_DISPLAY
+        std::strcpy(fpsOverlayBuf, "FPS 0");
+        fpsUpdateCounter = 0;
+        fpsAccumulatedMs = 0;
+#endif
     }
 
     Engine::Engine(const DisplayConfig& displayConfig, const InputConfig& inputConfig) 
         : renderer(displayConfig), inputManager(inputConfig), audioEngine(AudioConfig()), musicPlayer(audioEngine) {
         previousMillis = 0;
         deltaTime = 0;
+#ifdef PIXELROOT32_ENABLE_FPS_DISPLAY
+        std::strcpy(fpsOverlayBuf, "FPS 0");
+        fpsUpdateCounter = 0;
+        fpsAccumulatedMs = 0;
+#endif
     }
 
     Engine::Engine(const DisplayConfig& displayConfig) 
         : renderer(displayConfig), inputManager(InputConfig(0)), audioEngine(AudioConfig()), musicPlayer(audioEngine) {
         previousMillis = 0;
         deltaTime = 0;
+#ifdef PIXELROOT32_ENABLE_FPS_DISPLAY
+        std::strcpy(fpsOverlayBuf, "FPS 0");
+        fpsUpdateCounter = 0;
+        fpsAccumulatedMs = 0;
+#endif
     }
 
     Engine::~Engine() {}
@@ -106,6 +124,28 @@ namespace pixelroot32::core {
     void Engine::draw() {
         renderer.beginFrame();
         sceneManager.draw(renderer);
+#ifdef PIXELROOT32_ENABLE_FPS_DISPLAY
+        drawFpsOverlay(renderer);
+#endif
         renderer.endFrame();
     }
+
+#ifdef PIXELROOT32_ENABLE_FPS_DISPLAY
+    void Engine::drawFpsOverlay(Renderer& r) {
+        fpsAccumulatedMs += deltaTime;
+        if (++fpsUpdateCounter >= FPS_UPDATE_INTERVAL) {
+            unsigned int fps = 0;
+            if (fpsAccumulatedMs > 0) {
+                fps = (1000u * static_cast<unsigned int>(FPS_UPDATE_INTERVAL)) / static_cast<unsigned int>(fpsAccumulatedMs);
+                if (fps > 999) fps = 999;
+            }
+            std::snprintf(fpsOverlayBuf, sizeof(fpsOverlayBuf), "FPS %u", fps);
+            fpsUpdateCounter = 0;
+            fpsAccumulatedMs = 0;
+        }
+        int x = r.getWidth() - 48;
+        if (x < 0) x = 0;
+        r.drawText(fpsOverlayBuf, static_cast<int16_t>(x), 12, Color::Green, 1);
+    }
+#endif
 }